[koei][nx] support new romfs patch files
[gust][pc] add PKG and Management stubs

+++ Cethleann/Structure/INFO2.cs
@@ -3,8 +3,8 @@ namespace Cethleann.Structure
 #pragma warning disable 1591
     public struct INFO2
     {
-        public int INFO0Count { get; set; }
-        public int INFO1Count { get; set; }
+        public long INFO0Count { get; set; }
+        public long INFO1Count { get; set; }
     }
 #pragma warning restore 1591
 }
